fix blip-2 ci

+++ tests/models/blip_2/test_modeling_blip_2.py
@@ -799,11 +799,13 @@ class Blip2ModelIntegrationTest(unittest.TestCase):
 
     def test_inference_opt_batched_beam_search(self):
         processor = Blip2Processor.from_pretrained("Salesforce/blip2-opt-2.7b")
-        model = Blip2ForConditionalGeneration.from_pretrained("Salesforce/blip2-opt-2.7b").to(torch_device)
+        model = Blip2ForConditionalGeneration.from_pretrained(
+            "Salesforce/blip2-opt-2.7b", torch_dtype=torch.float16
+        ).to(torch_device)
 
         # prepare image
         image = prepare_img()
-        inputs = processor(images=[image, image], return_tensors="pt").to(torch_device)
+        inputs = processor(images=[image, image], return_tensors="pt").to(torch_device, dtype=torch.float16)
 
         predictions = model.generate(**inputs, num_beams=2)
 
@@ -844,11 +846,13 @@ class Blip2ModelIntegrationTest(unittest.TestCase):
 
     def test_inference_t5_batched_beam_search(self):
         processor = Blip2Processor.from_pretrained("Salesforce/blip2-flan-t5-xl")
-        model = Blip2ForConditionalGeneration.from_pretrained("Salesforce/blip2-flan-t5-xl").to(torch_device)
+        model = Blip2ForConditionalGeneration.from_pretrained(
+            "Salesforce/blip2-flan-t5-xl", torch_dtype=torch.float16
+        ).to(torch_device)
 
         # prepare image
         image = prepare_img()
-        inputs = processor(images=[image, image], return_tensors="pt").to(torch_device)
+        inputs = processor(images=[image, image], return_tensors="pt").to(torch_device, dtype=torch.float16)
 
         predictions = model.generate(**inputs, num_beams=2)
